STABLE14-viced-threadnum-return-cast-20090323
LICENSE IPL10

cast return type from threadNum


(cherry picked from commit ede1c215d16d394fcd0afb7aacd827680be56cf6)

+++ src/viced/viced.c
@@ -318,7 +318,7 @@ ResetCheckDescriptors(void)
 int
 threadNum(void)
 {
-    return pthread_getspecific(rx_thread_id_key);
+    return (int)pthread_getspecific(rx_thread_id_key);
 }
 #endif
